agent_recommendation_system_final copy/requirements.txt
pandas>=1.5.0
numpy>=1.21.0
chromadb>=0.4.0
openai>=1.0.0
langchain>=0.0.300
langchain-openai>=0.0.1
langchain-community>=0.0.1
sentence-transformers>=2.2.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
pydantic>=2.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
requests>=2.28.0
instructor>=1.0.0
streamlit
streamlit-option-menu 